stock-agent/requirements.txt
# ============================================================
# 选股Agent项目依赖
# Python 3.11+
# ============================================================

# ---- LangChain 核心 ----
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-community>=0.3.0,<0.4.0

# ---- LangGraph 多Agent编排 ----
langgraph>=0.2.0,<0.3.0

# ---- OpenAI 兼容客户端（用于 DeepSeek） ----
openai>=1.50.0,<2.0.0

# ---- A股数据源 ----
akshare>=1.14.0

# ---- 数据处理 ----
pandas>=2.1.0,<3.0.0
numpy>=1.26.0,<2.0.0

# ---- 环境变量管理 ----
python-dotenv>=1.0.0

# ---- HTTP 客户端（akshare 依赖） ----
requests>=2.31.0
httpx>=0.27.0

# ---- 异步支持 ----
aiohttp>=3.9.0

# ---- 数据验证 ----
pydantic>=2.5.0,<3.0.0

# ---- 工具依赖 ----
tenacity>=8.2.0        # 重试机制
tqdm>=4.66.0           # 进度条（批量分析时使用）

# ---- 开发和测试（可选） ----
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# black>=23.0.0
# isort>=5.12.0
